UKPE-1168 renamed /api/endpoints/updateHealth to /api/endpoints/health and done cleanup
