+++ HealthMonitoring.AcceptanceTests/HealthMonitoring.AcceptanceTests.csproj
@@ -178,7 +178,9 @@ xcopy  /E /Y /I $(SolutionDir)HealthMonitoring.Monitors.Nsb3\$(OutDir). $(Target
 xcopy  /E /Y /I $(SolutionDir)HealthMonitoring.SampleNsbHost\$(OutDir). $(TargetDir)sample
 
 xcopy  /Y /I $(ProjectDir)HealthMonitoring.SelfHost.exe.config $(TargetDir)api\
-xcopy  /Y /I $(ProjectDir)HealthMonitoring.Monitors.SelfHost.exe.config $(TargetDir)monitor\</PreBuildEvent>
+xcopy  /Y /I $(ProjectDir)HealthMonitoring.Monitors.SelfHost.exe.config $(TargetDir)monitor\
+xcopy  /Y /I $(ProjectDir)log4net.config $(TargetDir)api\
+xcopy  /Y /I $(ProjectDir)log4net.config $(TargetDir)monitor\</PreBuildEvent>
   </PropertyGroup>
   <Import Project="..\packages\System.Data.SQLite.Core.1.0.99.0\build\net451\System.Data.SQLite.Core.targets" Condition="Exists('..\packages\System.Data.SQLite.Core.1.0.99.0\build\net451\System.Data.SQLite.Core.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
